move env under installer
