+++ pkg/fleet/installer/default_packages_test.go
@@ -8,7 +8,7 @@ package installer
 import (
 	"testing"
 
-	"github.com/DataDog/datadog-agent/pkg/fleet/env"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/env"
 	"github.com/DataDog/datadog-agent/pkg/fleet/internal/oci"
 	"github.com/stretchr/testify/assert"
 )
