+++ pkg/fleet/internal/bootstrap/bootstrap.go
@@ -8,7 +8,8 @@ package bootstrap
 
 import (
 	"context"
-	"github.com/DataDog/datadog-agent/pkg/fleet/env"
+
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/env"
 )
 
 const (
