+++ pkg/fleet/internal/oci/download.go
@@ -33,7 +33,7 @@ import (
 	"golang.org/x/net/http2"
 	httptrace "gopkg.in/DataDog/dd-trace-go.v1/contrib/net/http"
 
-	"github.com/DataDog/datadog-agent/pkg/fleet/env"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/env"
 	"github.com/DataDog/datadog-agent/pkg/fleet/internal/tar"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
